retrieveImageData
We may wish to reuse this for the other editor.
